# backend/openapi.json
{"openapi": "3.1.0", "info": {"title": "Engram Backend API", "version": "0.1.0"}, "paths": {"/users": {"post": {"summary": "Sign In User", "operationId": "sign_in_user_users_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateUserRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/agents": {"get": {"summary": "Get User Agents", "operationId": "get_user_agents_agents_get", "parameters": [{"name": "authorization", "in": "header", "required": true, "schema": {"type": "string", "title": "Authorization"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ListAgentsResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "post": {"summary": "Create Agent", "operationId": "create_agent_agents_post", "parameters": [{"name": "authorization", "in": "header", "required": true, "schema": {"type": "string", "title": "Authorization"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateAgentRequest"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Agent"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/twitter/auth-url": {"post": {"summary": "Generate Twitter Auth Url", "operationId": "generate_twitter_auth_url_twitter_auth_url_post", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TwitterAuthUrlResponse"}}}}}}}, "/twitter/access-token": {"post": {"summary": "Sets Twitter Access Token", "operationId": "sets_twitter_access_token_twitter_access_token_post", "parameters": [{"name": "authorization", "in": "header", "required": true, "schema": {"type": "string", "title": "Authorization"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TwitterAccessTokenRequest"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/twitter/is-authenticated": {"get": {"summary": "Is Twitter Authenticated", "operationId": "is_twitter_authenticated_twitter_is_authenticated_get", "parameters": [{"name": "authorization", "in": "header", "required": true, "schema": {"type": "string", "title": "Authorization"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/seeds": {"get": {"summary": "Get Seeds By User", "operationId": "get_seeds_by_user_seeds_get", "parameters": [{"name": "authorization", "in": "header", "required": true, "schema": {"type": "string", "title": "Authorization"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ListSeedsResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"Agent": {"properties": {"id": {"type": "string", "format": "uuid", "title": "Id"}, "user_id": {"type": "string", "format": "uuid", "title": "User Id"}, "name": {"type": "string", "title": "Name"}}, "type": "object", "required": ["id", "user_id", "name"], "title": "Agent"}, "CreateAgentRequest": {"properties": {"name": {"type": "string", "title": "Name"}, "context": {"type": "string", "title": "Context"}, "seed_ids": {"items": {"type": "string", "format": "uuid"}, "type": "array", "title": "Seed Ids"}}, "type": "object", "required": ["name", "context", "seed_ids"], "title": "CreateAgentRequest"}, "CreateUserRequest": {"properties": {"email": {"type": "string", "title": "Email"}, "name": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Name"}, "clerk_id": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Clerk Id"}, "sign_in_method": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Sign In Method", "default": "CLERK"}}, "type": "object", "required": ["email"], "title": "CreateUserRequest"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "ListAgentsResponse": {"properties": {"agents": {"items": {"$ref": "#/components/schemas/Agent"}, "type": "array", "title": "Agents"}}, "type": "object", "required": ["agents"], "title": "ListAgentsResponse"}, "ListSeedsResponse": {"properties": {"seeds": {"items": {"$ref": "#/components/schemas/Seed"}, "type": "array", "title": "Seeds"}}, "type": "object", "required": ["seeds"], "title": "ListSeedsResponse"}, "OAuthInfo": {"properties": {"token_type": {"type": "string", "title": "Token Type", "default": "bearer"}, "access_token": {"type": "string", "title": "Access Token"}, "refresh_token": {"type": "string", "title": "Refresh Token"}, "expires_in": {"type": "integer", "title": "Expires In"}, "scope": {"items": {"type": "string"}, "type": "array", "title": "Scope"}, "expires_at": {"type": "number", "title": "Expires At"}}, "type": "object", "required": ["access_token", "refresh_token", "expires_in", "scope", "expires_at"], "title": "OAuthInfo"}, "Seed": {"properties": {"id": {"type": "string", "format": "uuid", "title": "Id"}, "user_id": {"type": "string", "format": "uuid", "title": "User Id"}, "seed_type": {"$ref": "#/components/schemas/SeedType"}, "auth_info": {"$ref": "#/components/schemas/OAuthInfo"}}, "type": "object", "required": ["id", "user_id", "seed_type", "auth_info"], "title": "Seed"}, "SeedType": {"type": "string", "enum": ["twitter"], "const": "twitter", "title": "SeedType"}, "TwitterAccessTokenRequest": {"properties": {"auth_response_url": {"type": "string", "title": "Auth Response Url"}}, "type": "object", "required": ["auth_response_url"], "title": "TwitterAccessTokenRequest"}, "TwitterAuthUrlResponse": {"properties": {"auth_url": {"type": "string", "title": "Auth Url"}}, "type": "object", "required": ["auth_url"], "title": "TwitterAuthUrlResponse"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}